requirements.txt
openai
anthropic
google-generativeai
pandas
python-dotenv
Pillow
